% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{r18r_example_translation}
\alias{r18r_example_translation}
\title{Return a translation}
\usage{
r18r_example_translation(language = translate_get_language("r18r.example"))
}
\description{
Return a translation
}
